LF-5298 Move TimeStrip component and stories
