+++ packages/webapp/src/components/TimeStrip/styles.module.scss
